REFACTOR: apply example file structure to app
