+++ Frontend/sopsc-mobile-app/src/components/user/Login.tsx
@@ -1,0 +1,99 @@
+import React, { useEffect, useState } from 'react';
+import { View, Text, Button, StyleSheet, TextInput, ActivityIndicator } from 'react-native';
+import { GoogleSignin, statusCodes } from '@react-native-google-signin/google-signin';
+import { useAuth } from '../../hooks/useAuth';
+
+interface LoginProps {
+  onLoginSuccess: (user: any) => void;
+}
+
+const Login = ({ onLoginSuccess }: LoginProps) => {
+  const [email, setEmail] = useState('');
+  const [password, setPassword] = useState('');
+
+  const { user, loading, signInEmail, signInGoogle } = useAuth();
+
+  useEffect(() => {
+    GoogleSignin.configure({
+      webClientId: process.env.EXPO_PUBLIC_GOOGLE_WEB_CLIENT_ID,
+      offlineAccess: true,
+    });
+  }, []);
+
+  useEffect(() => {
+    if (user) {
+      onLoginSuccess(user);
+    }
+  }, [user]);
+
+  const googleSignIn = async () => {
+    try {
+      await GoogleSignin.hasPlayServices();
+      const userInfo: any = await GoogleSignin.signIn();
+      await signInGoogle(userInfo.data?.idToken, userInfo.user?.name, userInfo.user?.email);
+    } catch (error: any) {
+      if (error.code === statusCodes.SIGN_IN_CANCELLED) {
+        alert('Sign in cancelled');
+      } else if (error.code === statusCodes.IN_PROGRESS) {
+        alert('Sign in in progress');
+      } else if (error.code === statusCodes.PLAY_SERVICES_NOT_AVAILABLE) {
+        alert('Google Play Services not available');
+      } else {
+        alert(`Unknown error: ${JSON.stringify(error)}`);
+      }
+    }
+  };
+
+  if (loading) {
+    return (
+      <View style={styles.container}>
+        <ActivityIndicator />
+      </View>
+    );
+  }
+
+  return (
+    <View style={styles.container}>
+      <Text style={styles.title}>SOPSC Sign In</Text>
+      <TextInput
+        style={styles.input}
+        placeholder="Email"
+        value={email}
+        onChangeText={setEmail}
+        autoCapitalize="none"
+      />
+      <TextInput
+        style={styles.input}
+        placeholder="Password"
+        value={password}
+        onChangeText={setPassword}
+        secureTextEntry
+      />
+      <Button title="Sign In" onPress={() => signInEmail(email, password)} />
+      <View style={{ height: 10 }} />
+      <Button title="Sign In with Google" onPress={googleSignIn} />
+    </View>
+  );
+};
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+    backgroundColor: '#fff',
+  },
+  title: {
+    fontSize: 20,
+    marginBottom: 20,
+  },
+  input: {
+    width: '80%',
+    padding: 10,
+    borderColor: '#ccc',
+    borderWidth: 1,
+    marginBottom: 10,
+  },
+});
+
+export default Login;
